update
Former-commit-id: 0e950e8e674431ee9df5d1e351c6f029879c1f6a

+++ user/llisiw/fatomoomp.c
@@ -196,7 +196,7 @@ void fatomo_lop(bool adj, bool add, int nx, int nr, float *x, float *r)
     }
 }
 
-void fatomo_ray(int **ray)
+void fatomo_ray(float **ray)
 /*< extract ray density >*/
 {
     int it, is;
@@ -206,10 +206,10 @@ void fatomo_ray(int **ray)
 #endif
     for (is=0; is < ns; is++) {
 	for (it=0; it <= nt; it++)
-	    ray[is][it] = 0;
+	    ray[is][it] = 0.;
 	
 	for (it=0; it < list[is][1]; it++) {
-	    ray[is][mask[is][it]] = 1;
+	    ray[is][mask[is][it]] = 1.;
 	}
 	
 	upgrad_ray(upgnum[is],upglist[is],ray[is]);
